Dont print char*

+++ tasks/bdev/src/bdev_runtime.cc
@@ -226,7 +226,7 @@ class Server : public Module {
       case BlockUrl::kFs: {
         HILOG(kInfo, "(node {}) Reading from FS, alloc={} off={} ptr={}",
               CHI_CLIENT->node_id_, task->data_.alloc_id_,
-              task->data_.off_.load(), data);
+              task->data_.off_.load(), (size_t)data);
         // ssize_t ret = pread(fd_, data, task->size_, task->off_);
         // if (ret == task->size_) {
         //   task->success_ = true;
